Add logo into email
