+++ src/Service/ImageService.php
@@ -31,6 +31,11 @@ class ImageService {
     }
   }
 
+  public function getLogo(): ?Image {
+    $publicFolder = $this->params->get('app.public_image');
+    return $this->loadImageFromPath('logo.png', "$publicFolder/logo.png");
+  }
+
   public function importLogo(UploadedFile $file): string {
     $publicFolder = $this->params->get('app.public_image');
     $this->createFolderIfNotExist($publicFolder);
@@ -59,7 +64,7 @@ class ImageService {
       'jpg',
       'JPG',
       'jpeg',
-      'JPEG'
+      'JPEG',
     ];
 
     $memberImage = $this->params->get('app.members_photos') . "/$licence";
@@ -69,5 +74,60 @@ class ImageService {
       }
     }
     return null;
+  }
+
+  public function loadImageFromProtectedPath(string $publicId): ?Image {
+    $path = $this->decodeEncodedUriId($publicId);
+
+    if (!$path || str_contains('./', $path)) {
+      return null;
+    }
+
+    // Image accessible to everyone logged
+    $imageFolder = $this->params->get('app.images');
+    return $this->loadImageFromPath($publicId, "$imageFolder/$path");
+  }
+
+  public function loadImageFromPublicPath(string $publicId): ?Image {
+    $path = $this->decodeEncodedUriId($publicId);
+
+    // Public resource is at root
+    if (!$path || str_contains('/', $path)) {
+      return null;
+    }
+
+    $imageFolder = $this->params->get('app.public_image');
+    return $this->loadImageFromPath($publicId, "$imageFolder/$path");
+  }
+
+  private function loadImageFromPath(string $publicId, string $path): ?Image {
+    if ($this->fs->exists($path)) {
+      $filename = explode("/", $path);
+      $filename = end($filename);
+
+      $image = new Image();
+      $image->setId($publicId)
+            ->setName($filename);
+
+      $this->setDataUri($path, $image);
+
+      return $image;
+    }
+    return null;
+  }
+
+  private function decodeEncodedUriId(string $encodedId): ?string {
+    if (!ctype_xdigit($encodedId)) return null;
+
+    return hex2bin($encodedId);
+  }
+
+  private function setDataUri($imagePath, Image $image): void {
+    $finfo = new \finfo(FILEINFO_MIME_TYPE);
+    $type = $finfo->file($imagePath);
+
+    $data = "data:$type;base64," . base64_encode(file_get_contents($imagePath));
+    $image->setMimeType($type)
+          ->setBase64($data);
   }
 }
